Disable segments
Extract segments handling into a separate module. Introduced config
value to enable segments.

Segments were not probably working for the last 6 month, the error
`Memory segment #0 has exceeded 100 KB length limit` constantly occurs,
even with a small amount of rooms.

Due to the 'new' isolated vms, the variable cache should be way more
reliable. This should be the main focus. Segments needs to be though
throw again, or the current system better understood and fixed.
