+++ src/require.js
@@ -3,6 +3,7 @@ require('logging');
 require('utils');
 require('config_brain_main');
 require('config_brain_memory');
+require('config_brain_segments');
 require('config_brain_nextroom');
 require('config_brain_quests');
 require('config_brain_squadmanager');
